Fix links to Travis build pages when Pro enabled
The links generated in the status bar were still resolving to 'travis-ci.org' even when Travis Pro was enabled. I found a spot where a url was generated without checking the atom.travis.pro setting, so always used the non-pro url. Added a check for the pro status to conditionally set url in a similar fashion as is used elsewhere in the code.